pyproject.toml
[project]
name = "volatility-forecasting-lab"
version = "0.1.0"
description = "Reproducible volatility forecasting research lab using public market data."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.2",
    "pyyaml>=6.0",
    "scikit-learn>=1.4",
    "statsmodels>=0.14",
    "tabulate>=0.9",
    "yfinance>=0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
